add debtors list

+++ app/src/main/java/com/example/partycheckapp/data/user/User.kt
@@ -3,8 +3,8 @@ package com.example.partycheckapp.data.user
 data class User (
     val id: Long,
     val name: String,
-    val phoneNumber: Int?,
-    val cardNumber: Int?,
+    val phoneNumber: Long?,
+    val cardNumber: Long?,
     val found: Boolean,
     val debitorList: ArrayList<Long>,
     val partyList: ArrayList<Long>,
